use ruby 2.4 in amazon linux 2
Install ruby 2.4 when using amazon linux 2: for older version of ruby, like 2.0, codedeploy has nasty bugs and memory leaks when deploying big packages.
Older versions of Ansible (like the one that comes in the default repository of amazon linux 2) have a bug and write "(kaboo)" instead of "2" for ansible_distribution_version, accomodate for this behaviour.